Extraced a sorted result set for search#filter
We lose about 10% of speed but gain 999999% readability
